+++ internal/service/history.go
@@ -9,7 +9,7 @@ import (
 
 type HistoryService interface {
 	GetHistory(ctx context.Context, pagination domain.Pagination) ([]domain.History, error)
-	DeleteOneHistory(ctx context.Context, postId int64, uid int64) error
+	DeleteOneHistory(ctx context.Context, postId uint, uid int64) error
 	DeleteAllHistory(ctx context.Context, uid int64) error
 }
 
@@ -36,7 +36,7 @@ func (h *historyService) GetHistory(ctx context.Context, pagination domain.Pagin
 
 	return history, nil
 }
-func (h *historyService) DeleteOneHistory(ctx context.Context, postId int64, uid int64) error {
+func (h *historyService) DeleteOneHistory(ctx context.Context, postId uint, uid int64) error {
 	if err := h.repo.DeleteOneHistory(ctx, postId, uid); err != nil {
 		return err
 	}
